+++ rollup.config.js
@@ -1,26 +1,37 @@
-import filesize from 'rollup-plugin-filesize';
+import fileSize from 'rollup-plugin-filesize';
 
-import filesizeBadger from './dist/index.esm.js';
+import {rollupPluginFilesizeBadger} from './dist/index.esm.js';
 
+/**
+* @external RollupOptions
+*/
+
+/**
+ * @param {PlainObject} cfg
+ * @param {"cjs"|"umd"|"esm"|"iife"} cfg.format
+ * @returns {RollupOptions}
+ */
 function getRollupObject ({
   format
 }) {
   return {
     plugins: [
-      filesize({
+      fileSize({
         showBeforeSizes: true,
-        postRender: filesizeBadger()
+        postRender: rollupPluginFilesizeBadger()
       })
     ],
-    external: ['fs', 'util'],
+    external: ['fs', 'util', 'path'],
     input: 'src/index.js',
     output: {
+      sourcemap: true,
       file: `dist/index.${format}.js`,
       format
     }
   };
 }
 
+// eslint-disable-next-line import/no-anonymous-default-export
 export default [
   getRollupObject({format: 'cjs'}),
   getRollupObject({format: 'esm'})